server/QR/requirements.txt
pandas==2.1.0
qrcode==7.4.2
Pillow==10.0.0 